move load validator in separate package
